+++ posthog-react-native/src/hooks/useNavigationTracker.ts
@@ -3,12 +3,17 @@ import { OptionalReactNativeNavigation } from '../optional/OptionalReactNativeNa
 import type { PostHog } from '../posthog-rn'
 import { PostHogAutocaptureNavigationTrackerOptions } from '../types'
 import { usePostHog } from './usePostHog'
+import { PostHogNavigationRef } from '../types'
 
 function _useNavigationTrackerDisabled(): void {
   return
 }
 
-function _useNavigationTracker(options?: PostHogAutocaptureNavigationTrackerOptions, client?: PostHog): void {
+function _useNavigationTracker(
+  options?: PostHogAutocaptureNavigationTrackerOptions,
+  navigationRef?: PostHogNavigationRef,
+  client?: PostHog
+): void {
   const contextClient = usePostHog()
   const posthog = client || contextClient
 
@@ -18,11 +23,35 @@ function _useNavigationTracker(options?: PostHogAutocaptureNavigationTrackerOpti
   }
 
   const routes = OptionalReactNativeNavigation.useNavigationState((state) => state?.routes)
-  const navigation = OptionalReactNativeNavigation.useNavigation()
+  // eslint-disable-next-line react-hooks/rules-of-hooks
+  const navigation = navigationRef || OptionalReactNativeNavigation.useNavigation()
 
   const trackRoute = useCallback((): void => {
+    if (!navigation) {
+      return
+    }
+
+    let currentRoute = undefined
+
     // NOTE: This method is not typed correctly but is available and takes care of parsing the router state correctly
-    const currentRoute = (navigation as any).getCurrentRoute()
+    try {
+      let isReady = false
+      try {
+        isReady = (navigation as any).isReady()
+      } catch (error) {
+        // keep compatibility with older versions of react-navigation
+        isReady = true
+      }
+
+      if (!isReady) {
+        return
+      }
+
+      currentRoute = (navigation as any).getCurrentRoute()
+    } catch (error) {
+      return
+    }
+
     if (!currentRoute) {
       return
     }