+++ posthog-web/CHANGELOG.md
@@ -1,8 +1,12 @@
 # Next
 
+# 3.5.0 – 2025-04-17
+
 ## Added
 
 1. chore: roll out new flag evaluation backend to majority of customers
+
+# 3.4.2 - 2025-02-27
 
 ## Added